+++ Samples~/Drag And Drop/Scripts/Editor/DragAndDrop.stories.cs
@@ -19,9 +19,9 @@ namespace Unity.AppUI.Sample.Editor
             {
                 var element = new VisualElement();
                 var tree = AssetDatabase.LoadAssetAtPath<VisualTreeAsset>(
-                        "Assets/Samples/App UI/1.0.0-pre.2/Drag And Drop/SampleResources/Drag And Drop/DragAndDropUI.uxml");
+                        "Assets/Samples/App UI/1.0.0-pre.3/Drag And Drop/SampleResources/Drag And Drop/DragAndDropUI.uxml");
                 var styleSheet = AssetDatabase.LoadAssetAtPath<ThemeStyleSheet>(
-                    "Assets/Samples/App UI/1.0.0-pre.2/Drag And Drop/SampleResources/Drag And Drop/DragAndDropTheme.tss");
+                    "Assets/Samples/App UI/1.0.0-pre.3/Drag And Drop/SampleResources/Drag And Drop/DragAndDropTheme.tss");
                 tree.CloneTree(element);
                 var root = element.Q<VisualElement>("main-root");
                 var desc = root.Q<Text>("dnd-desc");