+++ Documentation~/whats-new.md
@@ -2,7 +2,7 @@
 uid: whats-new
 ---
 
-# What's New in **1.0.0-pre.2**
+# What's New in **1.0.0-pre.3**
 
 This section contains information about new features, improvements, and issues fixed.
 
@@ -10,6 +10,10 @@ For a complete list of changes made, refer to the **Changelog** page.
 
 The main updates in this release include:
 
-### Added
+### Fixed
 
-- Added missing XML documentation parts for the public API of the package.
+- Fixed warning message about Z-axis scale in the `AnchorPopup` component.
+
+### Changed
+
+- The `Pressable` manipulator now uses the `keepPropagation` property also in its `PointerUpEvent` callback to avoid to propagate the event to the parent element.